Eusuite&DeploymentNando/apps/euadmin-backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-jose[cryptography]==3.3.0
argon2-cffi==23.1.0
python-multipart==0.0.6
httpx==0.26.0
sqlalchemy==2.0.25
pydantic[email]==2.5.3
pydantic-settings==2.1.0
kubernetes==28.1.0
python-dateutil==2.8.2
pyyaml==6.0.1
websockets==12.0
aiofiles==23.2.1
